+++ hebe/src/controllers/tanita.controller.ts
@@ -6,15 +6,19 @@ import { TanitaService } from '@services/index'
 import { TanitaStoreResDTO } from '@utils/tanita.dto'
 
 const TanitaController = async (app, opts) => {
-  app.post('/api/tanita/:clientId', { preValidation: [app.authGuard], schema: createTanitaSchema }, async (request: FastifyRequest, reply: FastifyReply): Promise<TanitaStoreResDTO> => {
-    const init = metricsInit()
+  app.post('/api/tanita/:clientId', { preValidation: [app.authGuard], schema: createTanitaSchema }, async (request: FastifyRequest, reply: FastifyReply): Promise<TanitaStoreResDTO | null> => {
+    const { role }: any = request.user
+    if (role === 'admin') {
+      const init = metricsInit()
 
-    const { clientId }: any = request.params
-    const body: any = request.body
-    const result: TanitaStoreResDTO = await TanitaService.store(body, clientId)
+      const { clientId }: any = request.params
+      const body: any = request.body
+      const result: TanitaStoreResDTO = await TanitaService.store(body, clientId)
 
-    metricsEnd(init)
-    return reply.send({ ...result })
+      metricsEnd(init)
+      return reply.send({ ...result })
+    }
+    return reply.send(null)
   })
 
   app.get('/api/tanita/:clientId', { preValidation: [app.authGuard], schema: FindAllSchema }, async (request: FastifyRequest, reply: FastifyReply): Promise<TanitaStoreResDTO[]> => {
@@ -28,34 +32,46 @@ const TanitaController = async (app, opts) => {
   })
 
   app.patch('/api/tanita/:id', { preValidation: [app.authGuard], schema: patchTanitaSchema }, async (request: FastifyRequest, reply: FastifyReply): Promise<null> => {
-    const init = metricsInit()
+    const { role }: any = request.user
+    if (role === 'admin') {
+      const init = metricsInit()
 
-    const body: any = request.body
-    const { id }: any = request.params
-    await TanitaService.patch(body, id)
+      const body: any = request.body
+      const { id }: any = request.params
+      await TanitaService.patch(body, id)
 
-    metricsEnd(init)
+      metricsEnd(init)
+      return reply.send(null)
+    }
     return reply.send(null)
   })
 
   app.delete('/api/tanita/:id', { preValidation: [app.authGuard], schema: IdParamsSchema }, async (request: FastifyRequest, reply: FastifyReply): Promise<null> => {
-    const init = metricsInit()
+    const { role }: any = request.user
+    if (role === 'admin') {
+      const init = metricsInit()
 
-    const { id }: any = request.params
-    await TanitaService.destroy(id)
+      const { id }: any = request.params
+      await TanitaService.destroy(id)
 
-    metricsEnd(init)
+      metricsEnd(init)
+      return reply.send(null)
+    }
     return reply.send(null)
   })
 
-  app.get('/api/tanita/detail/:id', { preValidation: [app.authGuard], schema: IdParamsSchema }, async (request: FastifyRequest, reply: FastifyReply): Promise<TanitaStoreResDTO> => {
-    const init = metricsInit()
+  app.get('/api/tanita/detail/:id', { preValidation: [app.authGuard], schema: IdParamsSchema }, async (request: FastifyRequest, reply: FastifyReply): Promise<TanitaStoreResDTO | null> => {
+    const { role }: any = request.user
+    if (role === 'admin') {
+      const init = metricsInit()
 
-    const { id }: any = request.params
-    const result: TanitaStoreResDTO = await TanitaService.findOne(id)
+      const { id }: any = request.params
+      const result: TanitaStoreResDTO = await TanitaService.findOne(id)
 
-    metricsEnd(init)
-    return reply.send(result)
+      metricsEnd(init)
+      return reply.send(result)
+    }
+    return reply.send(null)
   })
 }
